test(e2e): shared-ns templates, json-patch updates, helpers split, drift-spec fix
- resource_manager_integration.go: trim the empty-normalization comment to two lines.
- Place the integration templates (admin resources) in the shared namespace; workspace refs name
  jupyter-k8s-shared explicitly (an omitted ref namespace resolves to the workspace's own namespace,
  no fallback). Services stay in the workspace namespace (ns-drop resolves them there).
- Use JSON merge patches for in-place workspace updates instead of re-applying a whole object with the
  same name: patch-workspace-switch-cluster.json (switch serviceName) and
  patch-workspace-remove-integration.json (integrationTemplateRefs: []). Removes
  workspace-switch-cluster.yaml and workspace-no-integration.yaml; adds patchWorkspaceFromFile.
- Move the integration test utils into helpers_integration.go (qualified ginkgo/gomega imports per the
  other helpers_*.go files); the spec file keeps only the specs.
- Make the backing-service drift spec actually exercise frozen-replay: it drifted the Service and then
  asserted "nothing changed" over 30s, but nothing enqueued a reconcile in that window (the fixture
  Service is not owned/watched, and the only requeue after Available is the 5m integration-probe
  cadence), so the assertion passed even if replay were broken. Now force a reconcile after drifting
  (touchWorkspaceToForceReconcile) and assert the frozen CACHE_PORT stays 6379 -- the direct signal
  that the reconcile ignored the drifted Service. Drop the deployment.generation check: the controller
  copies all workspace annotations onto the pod template, so the reconcile-nudge annotation bumps
  generation independently of any re-render, making it an unreliable no-roll signal once a nudge is used.

Focused e2e: 16/16 green on an isolated kind cluster.
