+++ test/e2e/static/integration/patch-workspace-remove-integration.json
@@ -1,0 +1,5 @@
+{
+  "spec": {
+    "integrationTemplateRefs": []
+  }
+}
