--- test/e2e/static/integration/workspace-no-integration.yaml
@@ -1,18 +1,0 @@
-apiVersion: workspace.jupyter.org/v1alpha1
-kind: Workspace
-metadata:
-  name: workspace-with-integration
-  namespace: default
-spec:
-  # Same metadata.name as workspace-with-integration.yaml but with NO integrationTemplateRefs, so applying
-  # this UPDATES the existing workspace to drop its only integration. The controller must prune the frozen
-  # status entry and re-render the pod WITHOUT the sidecar (base-only), while the workspace stays Available.
-  desiredStatus: Running
-  image: jk8s-application-jupyter-uv:latest
-  resources:
-    requests:
-      cpu: 100m
-      memory: 128Mi
-    limits:
-      cpu: 500m
-      memory: 512Mi
