+++ test/e2e/static/integration/workspace-with-integration.yaml
@@ -7,7 +7,10 @@ spec:
   desiredStatus: Running
   image: jk8s-application-jupyter-uv:latest
   integrationTemplateRefs:
+      # The template is an admin resource in the shared namespace, so the ref names that namespace
+      # explicitly (an omitted namespace resolves to the workspace's own namespace only, no fallback).
     - name: service-integration
+      namespace: jupyter-k8s-shared
       parameters:
         - name: serviceName
           value: shared-cache
